update ui and be teacher

+++ backend/src/EnglishCentral.Shared/Common/PaginationHelpers/PagedResult.cs
@@ -1,14 +1,14 @@
 ﻿namespace EnglishCentral.Shared.Common.PaginationHelpers
 {
-    public record PagedResult<T>(IReadOnlyList<T> Items, long Page, long PageSize, int TotalItems, int TotalPages)
+    public record PagedResult<T>(IReadOnlyList<T> Items, int Page, int PageSize, int TotalItems, int TotalPages)
     {
         public bool HasPreviousPage => Page > 1;
         public bool HasNextPage => Page < TotalPages;
 
         public static PagedResult<T> Create(
             IReadOnlyList<T> items,
-            long page,
-            long pageSize,
+            int page,
+            int pageSize,
             int totalItems)
         {
             var totalPages = (int)Math.Ceiling(totalItems / (double)pageSize);
